requirements.txt
paddleocr
python-telegram-bot>=20.0
opencv-python
numpy
pillow
pandas
python-dotenv